requirements.txt
fastapi>=0.104.0
pydantic>=2.4.2
pydantic-ai>=0.5.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
jinja2>=3.1.2
sqlalchemy>=2.0.23
httpx>=0.25.1
python-multipart>=0.0.6
plotly>=5.18.0